Added buttons next to article content
