+++ client/src/components/Article/ArticleSocialMediaContainer.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import ButtonSmallRound from '../ButtonSmallRound'
+
+export default function ArticleSocialMediaContainer() {
+    return (
+        <div className='articleSocialMediaContainer'>
+            <ButtonSmallRound symbol="content_copy"/>
+            <ButtonSmallRound symbol="mail" />
+            <ButtonSmallRound symbol="share"/>
+
+        </div>
+    )
+}
