+++ client/src/components/Article/ArticleTitle.js
@@ -2,6 +2,8 @@ import React from 'react'
 
 export default function ArticleTitle() {
   return (
-    <div className='articleTitle'>ArticleTitle</div>
+    <div className='articleTitle'>
+      Our mission at Rentwallex
+    </div>
   )
 }
