+++ client/src/components/ButtonSmallRound.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function ButtonSmallRound(props) {
+    return (
+        <div className='buttonSmallRound'>
+            <span className="material-symbols-rounded">{props.symbol}</span>    
+        </div>
+    )
+}
