+++ client/src/pages/Mission.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Navbar from '../components/Navbar'
-import Article from '../components/Article'
+import Article from '../components/Article/Article'
 
 export default function Mission() {
   return (
@@ -8,7 +8,7 @@ export default function Mission() {
       {/* todo: change Navbar content if needed */}
       <Navbar />  
       <Article />
-      
+
       
     </div>
   )
